added .txt save to Node, removed web app/express , cleaned up dotnet layout
