+++ IQVIA-Project/Views/Home/About.cshtml
@@ -1,7 +1,10 @@
-﻿@{
+﻿
+@{
     ViewData["Title"] = "About";
 }
-<h2>@ViewData["Title"]</h2>
-<h3>@ViewData["Message"]</h3>
 
-<p>Use this area to provide additional information.</p>
+<div>
+    <p>Created by: Matthew Shim</p>
+    <p>Source Code: <a href="http://www.github.com/matthewshim-ms/IQVIA-API-Project">GitHub Repo</a></p>
+</div>
+
