+++ IQVIA-Project/wwwroot/css/site.css
@@ -40,3 +40,4 @@
     border: 1px solid black;
     border-radius: 5px;
 }
+
