--- Node/views/404.handlebars
@@ -1,1 +1,0 @@
-<h1>ERROR - HTTP:404 - Not Found</h1>
