--- Node/views/500.handlebars
@@ -1,1 +1,0 @@
-<h1>ERROR - HTTP:500 - Internal Server Error</h1>
