--- Node/views/layouts/main.handlebars
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8">
-	<title>IQVIA API Demo - JS</title>
-  <link rel="stylesheet" href="/css/styles.css">
-</head>
-<body>
-
-    <div class="container">
-        <fieldset>
-            <form action="/getTweets" method="GET">
-                <input type="submit" value="Get Tweets">
-            </form>
-        </fieldset>
-    </div>
-
-	{{{ body }}}
-
-<!-- Footer -->
-</body>
-
-<script type="text/javascript" src="/js/scripts.js"></script>
-</html>
